add some sort of customisable listener for the ratelimiter
this is just to give an idea of the functionality I think would be good. there should be some rate limiter listener that executes its handler before a rate limit occur, such that we have context of rate limits before they happen and can i.e send a warning about the rate limit. we should be able to specify the handler.

since the handler will be executed before the rate limit maybe the timing of that will cause complications I'm not sure.

just a thought